btrfs: Ensure that the home directory is a subvolume
Before this change, the btrfs graphdriver was not ensuring that the home
directory is a subvolume. Some Linux distributions were doing that on
their own in their packaging. This change ensures that on the driver
initialzation.

After this change, all [home]/subvolumes/* (by default:
/var/lib/docker/subvolumes) are going to be under the main home
subvolume.

When there is already an existing home directory which is not a
subvolume, it's going to be automatically converted to a subvolume.

Signed-off-by: Michal Rostecki <mrostecki@opensuse.org>